Move transform logic to opt
